saving timeline pseudo-reduxified
